o-ran-ssd-o1-control/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
python-dotenv==1.0.1
prometheus-client==0.21.0
pandas==2.2.3
numpy==2.1.2
requests>=2.31.0
